+++ backend/src/bin/worker/integrations.ts
@@ -7,17 +7,28 @@ import { NodeWorkerIntegrationProcessMessage } from '../../types/mq/nodeWorkerIn
 import { createRedisClient } from '../../utils/redis'
 import { NodeWorkerProcessWebhookMessage } from '../../types/mq/nodeWorkerProcessWebhookMessage'
 
+let integrationProcessorInstance: IntegrationProcessor
+
+async function getIntegrationProcessor(logger: Logger): Promise<IntegrationProcessor> {
+  if (integrationProcessorInstance) return integrationProcessorInstance
+
+  const options: IServiceOptions = {
+    ...(await SequelizeRepository.getDefaultIRepositoryOptions()),
+    log: logger,
+  }
+
+  const redisEmitter = await createRedisClient(true)
+
+  integrationProcessorInstance = new IntegrationProcessor(options, redisEmitter)
+
+  return integrationProcessorInstance
+}
+
 export const processIntegrationCheck = async (
   msg: NodeWorkerIntegrationCheckMessage,
   messageLogger: Logger,
 ): Promise<void> => {
-  const options = (await SequelizeRepository.getDefaultIRepositoryOptions()) as IServiceOptions
-  options.log = messageLogger
-
-  const redisEmitter = await createRedisClient(true)
-
-  const processor = new IntegrationProcessor(options, redisEmitter)
-
+  const processor = await getIntegrationProcessor(messageLogger)
   await processor.processCheck(msg.integrationType)
 }
 
@@ -25,13 +36,7 @@ export const processIntegration = async (
   msg: NodeWorkerIntegrationProcessMessage,
   messageLogger: Logger,
 ): Promise<void> => {
-  const options = (await SequelizeRepository.getDefaultIRepositoryOptions()) as IServiceOptions
-  options.log = messageLogger
-
-  const redisEmitter = await createRedisClient(true)
-
-  const processor = new IntegrationProcessor(options, redisEmitter)
-
+  const processor = await getIntegrationProcessor(messageLogger)
   await processor.process(msg)
 }
 
@@ -39,12 +44,6 @@ export const processWebhook = async (
   msg: NodeWorkerProcessWebhookMessage,
   messageLogger: Logger,
 ): Promise<void> => {
-  const options = (await SequelizeRepository.getDefaultIRepositoryOptions()) as IServiceOptions
-  options.log = messageLogger
-
-  const redisEmitter = await createRedisClient(true)
-
-  const processor = new IntegrationProcessor(options, redisEmitter)
-
-  await processor.processWebhook(msg.webhookId, msg.force)
+  const processor = await getIntegrationProcessor(messageLogger)
+  await processor.processWebhook(msg.webhookId, msg.force, msg.fireCrowdWebhooks)
 }