+++ backend/src/services/settingsService.ts
@@ -21,20 +21,21 @@ class SettingsService {
       throw new Error400(options.language, 'settings.activityTypes.errors.typeRequiredWhenCreating')
     }
 
-    const typeKey = data.type
+    const typeKey = data.type.toLowerCase()
+    const platformKey = platform.toLowerCase()
 
     const activityTypes = SettingsRepository.getActivityTypes(options)
 
-    if (!activityTypes.custom[platform]) {
-      activityTypes.custom[platform] = {}
+    if (!activityTypes.custom[platformKey]) {
+      activityTypes.custom[platformKey] = {}
     }
 
     // check key already exists
-    if (activityTypes.custom && activityTypes.custom[platform][typeKey]) {
+    if (activityTypes.custom && activityTypes.custom[platformKey][typeKey]) {
       return activityTypes
     }
 
-    activityTypes.custom[platform][typeKey] = {
+    activityTypes.custom[platformKey][typeKey] = {
       display: {
         default: data.type,
         short: data.type,