+++ frontend/src/integrations/zapier/config.js
@@ -10,4 +10,5 @@ export default {
     'https://www.seekpng.com/png/full/67-672759_zapiers-new-cli-tool-for-creating-apps-zapier.png',
   connectComponent: ZapierConnect,
   url: () => null,
+  chartColor: '#FF9676',
 };