+++ frontend/src/modules/automation/store/state.js
@@ -1,0 +1,7 @@
+export default () => ({
+  automations: [],
+  loadingAutomations: false,
+  filter: {
+    type: 'all',
+  },
+});