+++ frontend/src/modules/dashboard/components/activity/dashboard-activity-item.vue
@@ -46,7 +46,7 @@
               with-link
             />
             <div class="flex items-center">
-              <div>
+              <div v-if="platform">
                 <el-tooltip
                   effect="dark"
                   :content="platform.name"
@@ -59,6 +59,10 @@
                   />
                 </el-tooltip>
               </div>
+              <i
+                v-else
+                class="ri-radar-line text-base text-gray-400"
+              />
               <p class="flex text-2xs leading-4 pl-2">
                 <app-activity-message
                   :activity="activity"
@@ -99,7 +103,7 @@
               ><i
                  class="ri-lg ri-external-link-line mr-1"
                />
-                <span class="block">Open on {{ platform.name }}</span></a>
+                <span class="block">Open on {{ platform?.name || 'platform' }}</span></a>
             </div>
           </app-activity-content>
         </div>